add mostImpMatch, addPositions

+++ demo/src/main/java/com/example/demo/service/TournamentService.java
@@ -1,14 +1,10 @@
 package com.example.demo.service;
 
+import com.example.demo.dao.MatchDAO;
 import com.example.demo.dao.PlayerNodeDAO;
-import com.example.demo.model.Match;
-import com.example.demo.model.Player;
-import com.example.demo.model.PlayerNode;
-import com.example.demo.model.Tournament;
+import com.example.demo.model.*;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
 import java.util.stream.Collectors;
 
@@ -24,6 +20,7 @@ import org.slf4j.LoggerFactory;
 
 @Service
 
+//todo: aggiungere delete tournament
 public class TournamentService {
 
     private static final Logger logger = LoggerFactory.getLogger(TournamentService.class);
@@ -31,12 +28,14 @@ public class TournamentService {
     private final TournamentDAO tournamentDAO;
     private final PlayerDAO playerDAO;
     private final PlayerNodeDAO playerNodeDAO;
+    private final MatchDAO matchDAO;
 
     @Autowired
-    public TournamentService(TournamentDAO tournamentDAO, PlayerDAO playerDAO, PlayerNodeDAO playerNodeDAO) {
+    public TournamentService(MatchDAO matchDAO, TournamentDAO tournamentDAO, PlayerDAO playerDAO, PlayerNodeDAO playerNodeDAO) {
         this.tournamentDAO = tournamentDAO;
         this.playerDAO = playerDAO;
         this.playerNodeDAO = playerNodeDAO;
+        this.matchDAO = matchDAO;
     }
 
     public Tournament createTournament(Tournament tournament) throws RuntimeException {
@@ -45,7 +44,7 @@ public class TournamentService {
 
     public void deleteTournament(String tournamentId) {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        if(!Objects.equals(authentication.getName(), tournamentDAO.getTournament(tournamentId).getCreator())) {
+        if (!Objects.equals(authentication.getName(), tournamentDAO.getTournament(tournamentId).getCreator())) {
             throw new RuntimeException("Non sei il creatore del torneo");
         }
         tournamentDAO.deleteTournament(tournamentId);
@@ -78,14 +77,14 @@ public class TournamentService {
         return tournamentDAO.getActiveTournaments(playerElo);
     }
 
-    public void addMostImportantMatches(List<Match> matches, String tournamentId) {
-        try {
-            tournamentDAO.addMostImportantMatches(matches, tournamentId);
-        } catch (Exception e) {
-            logger.error("Errore durante l'aggiunta delle partite più importanti al torneo {}", tournamentId, e);
-            throw new RuntimeException("Errore nell'aggiungere le partite più importanti al torneo " + tournamentId);
-        }
-    }
+//    public void addMostImportantMatches(List<Match> matches, String tournamentId) {
+//        try {
+//            tournamentDAO.addMostImportantMatches(matches, tournamentId);
+//        } catch (Exception e) {
+//            logger.error("Errore durante l'aggiunta delle partite più importanti al torneo {}", tournamentId, e);
+//            throw new RuntimeException("Errore nell'aggiungere le partite più importanti al torneo " + tournamentId);
+//        }
+//    }
 
     public void joinTournament(String tournamentId) {
         try {
@@ -93,7 +92,9 @@ public class TournamentService {
             Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
             String playerUsername = authentication.getName();
 
-            tournamentDAO.joinTournament(tournamentId, playerUsername);
+            TournamentPlayer playerEntry = new TournamentPlayer(playerUsername, 0);
+
+            tournamentDAO.addPlayer(tournamentId, playerEntry);
         } catch (Exception e) {
             logger.error("Errore durante l'iscrizione al torneo {} da parte del giocatore {}", tournamentId,
                     e.getMessage(), e);
@@ -105,49 +106,62 @@ public class TournamentService {
         return tournamentDAO.getCreatedTournaments(creator);
     }
 
-    public void addMatch(String tournamentId, Match match) {
-        try {
-            tournamentDAO.addMatch(tournamentId, match);
-            logger.info("Match aggiunto con successo al torneo {}", tournamentId);
-        } catch (Exception e) {
-            logger.error("Errore durante l'aggiunta del match al torneo {}", tournamentId, e);
-            throw new RuntimeException("Errore nell'aggiungere il match al torneo " + tournamentId);
-        }
-    }
+//    public void addMatch(String tournamentId, Match match) {
+//        try {
+//            tournamentDAO.addMatch(tournamentId, match);
+//            logger.info("Match aggiunto con successo al torneo {}", tournamentId);
+//        } catch (Exception e) {
+//            logger.error("Errore durante l'aggiunta del match al torneo {}", tournamentId, e);
+//            throw new RuntimeException("Errore nell'aggiungere il match al torneo " + tournamentId);
+//        }
+//    }
 
+    // used by admin
     public void addPlayer(String tournamentId, String playerId) throws RuntimeException {
-            if(playerDAO.getPlayer(playerId) == null) {
-                throw new RuntimeException("Player not found");
-            }
+        if (playerDAO.getPlayer(playerId) == null) {
+            throw new RuntimeException("Player not found");
+        }
 
-            Tournament tournament = tournamentDAO.getTournament(tournamentId);
+        PlayerNode playerNode = playerNodeDAO.getPlayerById(playerId);
+        Tournament tournament = tournamentDAO.getTournament(tournamentId);
 
-            if(tournament.getIsClosed()) {
-                throw new RuntimeException("Torneo chiuso");
-            }
-            if(tournament.getPlayers().stream().anyMatch(playerMap -> playerMap.containsKey(playerId))) {
-                throw new RuntimeException("Giocatore già iscritto");
-            }
-            if(tournament.getPlayers().size()>=tournament.getMaxPlayers()) {
-                throw new RuntimeException("Torneo pieno");
-            }
+        if (playerNode.getElo() < tournament.getEloMin() || playerNode.getElo() > tournament.getEloMax()) {
+            throw new RuntimeException("Elo non valido");
+        }
 
-            Map<String, Integer> playerEntry = new HashMap<>();
-            playerEntry.put(playerId, 0);
-            tournamentDAO.addPlayer(tournamentId, playerEntry);
+        if (tournament.getIsClosed()) {
+            throw new RuntimeException("Torneo chiuso");
+        }
+        if (tournament.getPlayers().stream().anyMatch(p -> p.getUsername().equals(playerId))) {
+            throw new RuntimeException("Giocatore già iscritto");
+        }
+        if (tournament.getPlayers().size() >= tournament.getMaxPlayers()) {
+            throw new RuntimeException("Torneo pieno");
+        }
 
-            if(tournament.getPlayers().size() == tournament.getMaxPlayers()) {
-                tournamentDAO.closeTournament(tournamentId);
-            }
+        TournamentPlayer playerEntry = new TournamentPlayer(playerId, 0);
+        tournamentDAO.addPlayer(tournamentId, playerEntry);
+        playerDAO.addTournament(playerId, new PlayerTournament(tournamentId, tournament.getName(), tournament.getStartDate(), tournament.getEndDate(), 0));
 
-            logger.info("Giocatore {} aggiunto con successo al torneo {}", playerId, tournamentId);
+        if (tournament.getPlayers().size() + 1 == tournament.getMaxPlayers()) {
+            tournamentDAO.closeTournament(tournamentId);
+        }
+
+        logger.info("Giocatore {} aggiunto con successo al torneo {}", playerId, tournamentId);
     }
 
     public void removePlayer(String tournamentId, String playerId) {
         try {
-            //todo: capire perchè non rimuove il giocatore correttamente
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            String creator = tournamentDAO.getTournament(tournamentId).getCreator();
+
+            if (authentication.getName().equals(creator))
+                throw new RuntimeException("Non sei il creatore del torneo");
+
             tournamentDAO.removePlayer(tournamentId, playerId);
             tournamentDAO.openTournament(tournamentId);
+            playerDAO.removeTournament(tournamentId, playerId);
+
             logger.info("Giocatore {} rimosso con successo dal torneo {}", playerId, tournamentId);
         } catch (Exception e) {
             logger.error("Errore durante la rimozione del giocatore {} dal torneo {}", playerId, tournamentId, e);
@@ -156,4 +170,35 @@ public class TournamentService {
         }
     }
 
+    public void updatePositions(List<TournamentPlayer> tournamentPlayers, String tournamentId) {
+        try {
+            Tournament tournament = tournamentDAO.getTournament(tournamentId);
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            String creator = tournament.getCreator();
+
+            if (!Objects.equals(authentication.getName(), creator))
+                throw new RuntimeException("Non sei il creatore del torneo");
+
+            if (!tournament.getIsClosed())
+                throw new RuntimeException("Torneo ancora non chiuso");
+
+            tournamentDAO.updatePositions(tournamentPlayers, tournamentId);
+            playerDAO.updateTournamentPositions(tournamentPlayers, tournamentId);
+            logger.info("Posizioni aggiornate con successo per il torneo {}", tournamentId);
+        } catch (Exception e) {
+            logger.error("Errore durante l'aggiornamento delle posizioni per il torneo {}", tournamentId, e);
+            throw new RuntimeException("Errore nell'aggiornare le posizioni per il torneo " + tournamentId);
+        }
+    }
+
+    public void addMostImportantMatches(String tournamentId, List<TournamentMatch> matches) {
+        try {
+            tournamentDAO.addMatchToTournament(tournamentId, matches);
+            matchDAO.addMatches(matches.stream().map(TournamentMatch::getMatch).collect(Collectors.toList()));
+            logger.info("Partite aggiunte con successo al torneo {}", tournamentId);
+        } catch (Exception e) {
+            logger.error("Errore durante l'aggiunta delle partite al torneo {}", tournamentId, e);
+            throw new RuntimeException("Errore nell'aggiungere le partite al torneo " + tournamentId);
+        }
+    }
 }
